+++ src/entities/users/model/types/index.ts
@@ -1,0 +1,2 @@
+export * from './data-type';
+export * from './api-type';
