+++ src/features/users/index.ts
@@ -1,2 +1,2 @@
-export * from './api';
+export * from './ui';
 export * from './model';
