+++ src/features/users/ui/index.ts
@@ -1,0 +1,1 @@
+export { UserModal } from './UserModal';
