+++ src/pages/PostsManagerPage.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react';
 import {
   AddCommentDialog,
   AddPostButton,
@@ -6,52 +5,17 @@ import {
   EditCommentDialog,
   PostDialog,
   PostsFilter,
+  PostsPagination,
   PostsTable,
 } from '@/features/posts';
+import { UserModal, useUserModal } from '@/features/users';
 import { EditPostDialog } from '@/features/posts';
-import { useLimit, usePosts } from '@/entities/posts';
-import { useSkip } from '@/entities/posts';
-import {
-  Button,
-  Card,
-  CardContent,
-  CardHeader,
-  CardTitle,
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/shared/ui';
+import { Card, CardContent, CardHeader, CardTitle } from '@/shared/ui';
 
 const PostsManager = () => {
   // 상태 관리
 
-  const [skip, setSkip] = useSkip();
-
-  const [limit, setLimit] = useLimit();
-
-  const [showUserModal, setShowUserModal] = useState(false);
-  const [selectedUser, setSelectedUser] = useState(null);
-
-  const { data: postsData } = usePosts();
-  const total = postsData?.total ?? 0;
-
-  // 사용자 모달 열기
-  const openUserModal = async (user: { id: number }) => {
-    try {
-      const response = await fetch(`/api/users/${user.id}`);
-      const userData = await response.json();
-      setSelectedUser(userData);
-      setShowUserModal(true);
-    } catch (error) {
-      console.error('사용자 정보 가져오기 오류:', error);
-    }
-  };
+  const { open: openUserModal } = useUserModal();
 
   return (
     <Card className="w-full max-w-6xl mx-auto">
@@ -70,42 +34,10 @@ const PostsManager = () => {
           <PostsFilter />
 
           {/* 게시물 테이블 */}
-          <PostsTable onClickOpenUserModal={openUserModal} />
+          <PostsTable onClickOpenUserModal={(user) => openUserModal(user.id)} />
 
           {/* 페이지네이션 */}
-          <div className="flex justify-between items-center">
-            <div className="flex items-center gap-2">
-              <span>표시</span>
-              <Select
-                value={limit.toString()}
-                onValueChange={(value) => setLimit(Number(value))}
-              >
-                <SelectTrigger className="w-[180px]">
-                  <SelectValue placeholder="10" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="10">10</SelectItem>
-                  <SelectItem value="20">20</SelectItem>
-                  <SelectItem value="30">30</SelectItem>
-                </SelectContent>
-              </Select>
-              <span>항목</span>
-            </div>
-            <div className="flex gap-2">
-              <Button
-                disabled={skip === 0}
-                onClick={() => setSkip(Math.max(0, skip - limit))}
-              >
-                이전
-              </Button>
-              <Button
-                disabled={skip + limit >= total}
-                onClick={() => setSkip(skip + limit)}
-              >
-                다음
-              </Button>
-            </div>
-          </div>
+          <PostsPagination />
         </div>
       </CardContent>
 
@@ -125,46 +57,7 @@ const PostsManager = () => {
       <PostDialog />
 
       {/* 사용자 모달 */}
-      <Dialog open={showUserModal} onOpenChange={setShowUserModal}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>사용자 정보</DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4">
-            <img
-              src={selectedUser?.image}
-              alt={selectedUser?.username}
-              className="w-24 h-24 rounded-full mx-auto"
-            />
-            <h3 className="text-xl font-semibold text-center">
-              {selectedUser?.username}
-            </h3>
-            <div className="space-y-2">
-              <p>
-                <strong>이름:</strong> {selectedUser?.firstName}{' '}
-                {selectedUser?.lastName}
-              </p>
-              <p>
-                <strong>나이:</strong> {selectedUser?.age}
-              </p>
-              <p>
-                <strong>이메일:</strong> {selectedUser?.email}
-              </p>
-              <p>
-                <strong>전화번호:</strong> {selectedUser?.phone}
-              </p>
-              <p>
-                <strong>주소:</strong> {selectedUser?.address?.address},{' '}
-                {selectedUser?.address?.city}, {selectedUser?.address?.state}
-              </p>
-              <p>
-                <strong>직장:</strong> {selectedUser?.company?.name} -{' '}
-                {selectedUser?.company?.title}
-              </p>
-            </div>
-          </div>
-        </DialogContent>
-      </Dialog>
+      <UserModal />
     </Card>
   );
 };
